extractor/requirements.txt
requests==2.32.3
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1